requirements.txt
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.11.5
click==6.7
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Pillow==5.3.0
pycparser==2.18
pyparsing==2.4.5
six==1.11.0
SQLAlchemy==1.2.7
svgwrite==1.3.1
Werkzeug==0.14.1
WTForms==2.1
